requirements.txt
numpy
faiss-cpu
sentence-transformers
python-docx
pymupdf
chardet
